+++ ktlint-cli-reporter-baseline/src/main/kotlin/com/pinterest/ktlint/cli/reporter/baseline/BaselineReporterProvider.kt
@@ -5,6 +5,7 @@ import java.io.PrintStream
 
 public class BaselineReporterProvider : ReporterProviderV2<BaselineReporter> {
     override val id: String = "baseline"
+
     override fun get(
         out: PrintStream,
         opt: Map<String, String>,