+++ ktlint-cli-reporter-json/src/main/kotlin/com/pinterest/ktlint/cli/reporter/json/JsonReporterProvider.kt
@@ -5,6 +5,7 @@ import java.io.PrintStream
 
 public class JsonReporterProvider : ReporterProviderV2<JsonReporter> {
     override val id: String = "json"
+
     override fun get(
         out: PrintStream,
         opt: Map<String, String>,