+++ ktlint-rule-engine/src/main/kotlin/com/pinterest/ktlint/rule/engine/api/KtLintRuleEngine.kt
@@ -25,6 +25,7 @@ import org.ec4j.core.Resource
 import org.ec4j.core.model.PropertyType.EndOfLineValue.crlf
 import org.ec4j.core.model.PropertyType.EndOfLineValue.lf
 import org.jetbrains.kotlin.com.intellij.lang.FileASTNode
+import org.jetbrains.kotlin.utils.addToStdlib.ifFalse
 import java.nio.charset.StandardCharsets
 import java.nio.file.FileSystem
 import java.nio.file.FileSystems
@@ -49,6 +50,14 @@ public class KtLintRuleEngine(
      * property being set in any other way.
      */
     public val editorConfigOverride: EditorConfigOverride = EMPTY_EDITOR_CONFIG_OVERRIDE,
+    /**
+     * Temporary flag to indicate that kotlin embeddable compiler is to be executed with (default) or without the extension point
+     * 'org.jetbrains.kotlin.com.intellij.treeCopyHandler'. This extension point is not (yet) supported in the preview of Kotlin 1.9. Some
+     * rules might no longer work and throw exceptions at runtime.
+     * It is unclear whether the extension point will be supported. Disabling this flag on the Kotlin 1.8 compiler has the same effect. As
+     * of that it can be used to assess the impact, and to fix rules before release of the Ktlint version which will be based on Kotlin 1.9.
+     */
+    public val enableKotlinCompilerExtensionPoint: Boolean = true,
     /**
      * **For internal use only**: indicates that linting was invoked from KtLint CLI tool. It enables some internals workarounds for Kotlin
      * Compiler initialization. This property is likely to be removed in any of next versions without further notice.
@@ -152,6 +161,13 @@ public class KtLintRuleEngine(
                             canBeAutoCorrected,
                         ),
                     )
+                    // In trace mode report the violation immediately. The order in which violations are actually found might be different
+                    // from the order in which they are reported. For debugging purposes it cn be helpful to know the exact order in which
+                    // violations are being solved.
+                    LOGGER.trace {
+                        "Format violation: ${code.fileNameOrStdin()}:$line:$col: $errorMessage (${rule.ruleId})" +
+                            canBeAutoCorrected.ifFalse { " [cannot be autocorrected]" }
+                    }
                 }
             }
         if (tripped) {