+++ ktlint-rule-engine/src/test/kotlin/com/pinterest/ktlint/rule/engine/api/KtLintTest.kt
@@ -503,7 +503,7 @@ private class AutoCorrectErrorRule : Rule(
                 STRING_VALUE_TO_BE_AUTOCORRECTED -> {
                     emit(node.startOffset, ERROR_MESSAGE_CAN_BE_AUTOCORRECTED, true)
                     if (autoCorrect) {
-                        (node as LeafElement).replaceWithText(STRING_VALUE_AFTER_AUTOCORRECT)
+                        (node as LeafElement).rawReplaceWithText(STRING_VALUE_AFTER_AUTOCORRECT)
                     }
                 }
                 STRING_VALUE_NOT_TO_BE_CORRECTED ->
@@ -609,6 +609,7 @@ private data class RuleExecutionCall(
     val classIdentifier: String? = null,
 ) {
     enum class RuleMethod { BEFORE_FIRST, BEFORE_CHILDREN, VISIT, AFTER_CHILDREN, AFTER_LAST }
+
     enum class VisitNodeType { ROOT, CHILD }
 }