+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/rules/ParameterListWrappingRule.kt
@@ -105,7 +105,7 @@ public class ParameterListWrappingRule :
                             true,
                         )
                         if (autoCorrect) {
-                            lpar.upsertWhitespaceAfterMe("\n${indentConfig.indent}")
+                            lpar.upsertWhitespaceAfterMe(indentConfig.childIndentOf(node))
                         }
                     }
                 nullableType
@@ -118,7 +118,7 @@ public class ParameterListWrappingRule :
                             true,
                         )
                         if (autoCorrect) {
-                            rpar.upsertWhitespaceBeforeMe("\n")
+                            rpar.upsertWhitespaceBeforeMe(indentConfig.parentIndentOf(node))
                         }
                     }
             }
@@ -186,15 +186,13 @@ public class ParameterListWrappingRule :
         emit: (offset: Int, errorMessage: String, canBeAutoCorrected: Boolean) -> Unit,
         autoCorrect: Boolean,
     ) {
-        val newIndentLevel = getNewIndentLevel(node)
         node
             .children()
-            .forEach { child -> wrapParameterInList(newIndentLevel, child, emit, autoCorrect) }
+            .forEach { child -> wrapParameterInList(child, emit, autoCorrect) }
     }
 
-    private fun getNewIndentLevel(node: ASTNode): Int {
-        val currentIndentLevel = indentConfig.indentLevelFrom(node.indent(false))
-        return when {
+    private fun intendedIndent(child: ASTNode): String =
+        when {
             // IDEA quirk:
             // fun <
             //     T,
@@ -209,19 +207,28 @@ public class ParameterListWrappingRule :
             //         param1: T
             //         param2: R
             //     )
-            currentIndentLevel > 0 && node.hasTypeParameterListInFront() -> currentIndentLevel - 1
+            child.treeParent.hasTypeParameterListInFront() -> -1
 
-            else -> currentIndentLevel
+            else -> 0
+        }.let {
+            if (child.elementType == VALUE_PARAMETER) {
+                it + 1
+            } else {
+                it
+            }
+        }.let { indentLevelFix ->
+            val indentLevel =
+                indentConfig
+                    .indentLevelFrom(child.treeParent.indent(false))
+                    .plus(indentLevelFix)
+            "\n" + indentConfig.indent.repeat(indentLevel)
         }
-    }
 
     private fun wrapParameterInList(
-        newIndentLevel: Int,
         child: ASTNode,
         emit: (offset: Int, errorMessage: String, canBeAutoCorrected: Boolean) -> Unit,
         autoCorrect: Boolean,
     ) {
-        val indent = "\n" + indentConfig.indent.repeat(newIndentLevel)
         when (child.elementType) {
             LPAR -> {
                 val prevLeaf = child.prevLeaf()
@@ -241,12 +248,7 @@ public class ParameterListWrappingRule :
                 // ... LPAR
                 // <line indent + indentSize> VALUE_PARAMETER...
                 // <line indent> RPAR
-                val intendedIndent =
-                    if (child.elementType == VALUE_PARAMETER) {
-                        indent + indentConfig.indent
-                    } else {
-                        indent
-                    }
+                val intendedIndent = intendedIndent(child)
                 val prevLeaf = child.prevLeaf()
                 if (prevLeaf is PsiWhiteSpace) {
                     if (prevLeaf.getText().contains("\n")) {